Refectored the utilities package and moved model utilities to Model and view utilities to View
