+++ core/src/com/mygdx/chalmersdefense/model/modelUtilities/IGameTimer.java
@@ -1,4 +1,4 @@
-package com.mygdx.chalmersdefense.utilities;
+package com.mygdx.chalmersdefense.model.modelUtilities;
 
 /**
  * @author Joel Båtsman Hilmersson
