+++ core/src/com/mygdx/chalmersdefense/model/projectiles/MatrixArea.java
@@ -1,6 +1,6 @@
 package com.mygdx.chalmersdefense.model.projectiles;
 
-import com.mygdx.chalmersdefense.utilities.CountDownTimer;
+import com.mygdx.chalmersdefense.model.modelUtilities.CountDownTimer;
 
 /**
  * @author Joel Båtsman Hilmersson
