+++ core/src/com/mygdx/chalmersdefense/model/towers/TowerFactory.java
@@ -3,7 +3,7 @@ package com.mygdx.chalmersdefense.model.towers;
 
 import com.mygdx.chalmersdefense.model.Player;
 import com.mygdx.chalmersdefense.model.projectiles.IProjectile;
-import com.mygdx.chalmersdefense.utilities.PathRectangle;
+import com.mygdx.chalmersdefense.model.modelUtilities.PathRectangle;
 
 import java.util.List;
