+++ desktop/src/test/java/testUtilities/TestPathRectangle.java
@@ -1,7 +1,7 @@
 package testUtilities;
 
 
-import com.mygdx.chalmersdefense.utilities.PathRectangle;
+import com.mygdx.chalmersdefense.model.modelUtilities.PathRectangle;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
